Moving BuildSignPodManager from the `pod` to the `buildsign` package.
BuildSignPodManager interface is used only by the build and sign packages.
Since we are moving into buildsign package, it make sense to move that
interface there, especially as it is a porting layer
